+++ code/utils/trainer_qa.py
@@ -1,0 +1,177 @@
+# coding=utf-8
+# Copyright 2020 The HuggingFace Team All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+Question-Answering task와 관련된 'Trainer'의 subclass 코드 입니다.
+"""
+
+from transformers import Trainer, is_datasets_available, is_torch_tpu_available
+from transformers.trainer_utils import denumpify_detensorize
+
+
+if is_datasets_available():
+    import datasets
+
+if is_torch_tpu_available():
+    import torch_xla.core.xla_model as xm
+    import torch_xla.debug.metrics as met
+
+# Huggingface의 Trainer를 상속받아 QuestionAnswering을 위한 Trainer를 생성합니다.
+class QuestionAnsweringTrainer(Trainer):
+    """
+        Inherit Huggingface Trainer and create Trainer for QuestionAnswering.
+
+        Args:
+            eval_examples (:obj:`Dataset`):
+                The dataset to use 'post_process_function' function.
+            post_process_function (:obj:``):
+                The function that post-processes the prediction value of the qa model.
+
+    """
+    def __init__(self, *args, eval_examples=None, post_process_function=None, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.eval_examples = eval_examples
+        self.post_process_function = post_process_function
+
+    def evaluate(self, eval_dataset=None, eval_examples=None, ignore_keys=None, metric_key_prefix="eval"):
+        """
+        Run evaluation and returns metrics.
+
+        Args:
+            eval_dataset (:obj:`Dataset`):
+                The dataset to use for evaluation.
+            eval_examples (:obj:`Dataset`):
+                The dataset used for post-processing.
+            ignore_keys (:obj:`Lst[str]`):
+                A list of keys in the output of your model that should be ignored when gathering predictions.
+            metric_key_prefix (:obj:`str`, defaults to :obj:`"eval"`):
+                An optional prefix to be used as the metrics key prefix.
+
+        Returns:
+            metrics (dict):
+                Containing the potential metrics.
+        """
+        self._memory_tracker.start()
+
+        eval_dataset = self.eval_dataset if eval_dataset is None else eval_dataset
+        eval_dataloader = self.get_eval_dataloader(eval_dataset)
+        eval_examples = self.eval_examples if eval_examples is None else eval_examples
+
+        # 일시적으로 metric computation를 불가능하게 한 상태이며, 해당 코드에서는 loop 내에서 metric 계산을 수행합니다.
+        compute_metrics = self.compute_metrics
+        self.compute_metrics = None
+        try:
+            output = self.evaluation_loop(
+                eval_dataloader,
+                description="Evaluation",
+                # metric이 없으면 예측값을 모으는 이유가 없으므로 아래의 코드를 따르게 됩니다.
+                # self.args.prediction_loss_only
+                prediction_loss_only=True if compute_metrics is None else None,
+                ignore_keys=ignore_keys,
+            )
+        finally:
+            self.compute_metrics = compute_metrics
+
+        if isinstance(eval_dataset, datasets.Dataset):
+            eval_dataset.set_format(
+                type=eval_dataset.format["type"],
+                columns=list(eval_dataset.features.keys()),
+            )
+
+        if self.post_process_function is not None and self.compute_metrics is not None:
+            eval_preds = self.post_process_function(
+                eval_examples, eval_dataset, output.predictions, self.args
+            )
+            metrics = self.compute_metrics(eval_preds)
+
+            # To be JSON-serializable, we need to remove numpy types or zero-d tensors
+            metrics = denumpify_detensorize(metrics)
+
+            # Prefix all keys with metric_key_prefix + '_'
+            for key in list(metrics.keys()):
+                if not key.startswith(f"{metric_key_prefix}_"):
+                    if key == f"exact_match":
+                        metrics[f"{metric_key_prefix}_em"] = metrics.pop(key)
+                    else:    
+                        metrics[f"{metric_key_prefix}_{key}"] = metrics.pop(key)
+
+            self.log(metrics)
+        else:
+            metrics = {}
+
+        if self.args.tpu_metrics_debug or self.args.debug:
+            # tpu-comment: PyTorch/XLA에 대한 Logging debug metrics (compile, execute times, ops, etc.)
+            xm.master_print(met.metrics_report())
+
+        self.control = self.callback_handler.on_evaluate(
+            self.args, self.state, self.control, metrics
+        )
+
+        self._memory_tracker.stop_and_update_metrics(metrics)
+
+        return metrics
+
+    def predict(self, test_dataset, test_examples, ignore_keys=None):
+        """
+        Run prediction and returns predictions
+
+        Args:
+            test_dataset (:obj:`Dataset`):
+                The dataset to run the predictions on.
+            test_examples (:obj:`Dataset`):
+                The dataset used for post-processing.
+            ignore_keys(:obj:`Lst[str]`):
+                A list of keys in the output of your model that should be ignored when gathering predictions.
+
+        Returns:
+            If your compute_metrics or compute_metrics is not None, output is equal to 'evaluate' function output.
+            output (dict):
+                Containing the potential metrics.
+            but if not, make predictions using 'post_process_function' function. 
+            predictions(:obj:`Callable[EvalPrediction]`):
+                The results of post_process_function.
+
+        """
+        test_dataloader = self.get_test_dataloader(test_dataset)
+
+        # 일시적으로 metric computation를 불가능하게 한 상태이며, 해당 코드에서는 loop 내에서 metric 계산을 수행합니다.
+        # evaluate 함수와 동일하게 구성되어있습니다
+        compute_metrics = self.compute_metrics
+        self.compute_metrics = None
+        try:
+            output = self.evaluation_loop(
+                test_dataloader,
+                description="Evaluation",
+                # metric이 없으면 예측값을 모으는 이유가 없으므로 아래의 코드를 따르게 됩니다.
+                # self.args.prediction_loss_only
+                prediction_loss_only=True if compute_metrics is None else None,
+                ignore_keys=ignore_keys,
+            )
+        finally:
+            self.compute_metrics = compute_metrics
+
+        if self.post_process_function is None or self.compute_metrics is None:
+            return output
+
+        if isinstance(test_dataset, datasets.Dataset):
+            test_dataset.set_format(
+                type=test_dataset.format["type"],
+                columns=list(test_dataset.features.keys()),
+            )
+
+        predictions = self.post_process_function(
+            test_examples, test_dataset, output.predictions, self.args
+        )
+
+        return predictions
